feat: images directory, index.html and style.css
Created an images directory and added the photos
I would like to use. Also create an index.html
and used the genrated boiler plate as well as
started to make divs for elements in the header.
Created a style.css with a font in the body and
put the link code into index.html.